@@ -784,6 +784,7 @@ end
   #   impersonationAuthentication - ImpersonationAuthenticationType
   #   fingerPrint - FingerPrintType
   #   mobileDeviceId - SOAP::SOAPString
+  #   accessToken - SOAP::SOAPString
   class MerchantAuthenticationType
     include ROXML
     xml_accessor :name
@@ -793,8 +794,9 @@ end
     xml_accessor :impersonationAuthentication, :as => ImpersonationAuthenticationType
     xml_accessor :fingerPrint, :as => FingerPrintType
     xml_accessor :mobileDeviceId
+	xml_accessor :accessToken
   
-    def initialize(name = nil, transactionKey = nil, sessionToken = nil, password = nil, impersonationAuthentication = nil, fingerPrint = nil, mobileDeviceId = nil)
+    def initialize(name = nil, transactionKey = nil, sessionToken = nil, password = nil, impersonationAuthentication = nil, fingerPrint = nil, mobileDeviceId = nil, accessToken = nil)
       @name = name
       @transactionKey = transactionKey
       @sessionToken = sessionToken
@@ -802,6 +804,7 @@ end
       @impersonationAuthentication = impersonationAuthentication
       @fingerPrint = fingerPrint
       @mobileDeviceId = mobileDeviceId
+      @accessToken = accessToken
     end
   end
   
@@ -1678,6 +1681,8 @@ end
   #   mobileDeviceId - SOAP::SOAPString
   #   subscription - SubscriptionPaymentType
   #   hasReturnedItems - SOAP::SOAPBoolean
+  #   fraudInformation - FraudInformationType
+  #   profile - CustomerProfileIdType
   class TransactionSummaryType
     include ROXML
     xml_accessor :transId
@@ -1693,11 +1698,12 @@ end
     xml_accessor :marketType
     xml_accessor :product
     xml_accessor :mobileDeviceId
-    xml_accessor :subscription
+    xml_accessor :subscription, :as => SubscriptionPaymentType
     xml_accessor :hasReturnedItems
     xml_accessor :fraudInformation
+	xml_accessor :profile, :as => CustomerProfileIdType
   
-    def initialize(transId = nil, submitTimeUTC = nil, submitTimeLocal = nil, transactionStatus = nil, invoiceNumber = nil, firstName = nil, lastName = nil, accountType = nil, accountNumber = nil, settleAmount = nil, marketType = nil, product = nil, mobileDeviceId = nil, subscription = nil, hasReturnedItems = nil, fraudInformation = nil)
+    def initialize(transId = nil, submitTimeUTC = nil, submitTimeLocal = nil, transactionStatus = nil, invoiceNumber = nil, firstName = nil, lastName = nil, accountType = nil, accountNumber = nil, settleAmount = nil, marketType = nil, product = nil, mobileDeviceId = nil, subscription = nil, hasReturnedItems = nil, fraudInformation = nil, profile = nil)
       @transId = transId
       @submitTimeUTC = submitTimeUTC
       @submitTimeLocal = submitTimeLocal
@@ -1714,6 +1720,7 @@ end
       @subscription = subscription
       @hasReturnedItems = hasReturnedItems
       @fraudInformation = fraudInformation
+	  @profile = profile
     end
   end
 
@@ -1926,6 +1933,7 @@ end
   #   returnedItems - ArrayOfReturnedItem
   #   solution - SolutionType
   #   emvDetails - TransactionDetailsType::EmvDetails
+  #   profile - CustomerProfileIdType
   class TransactionDetailsType
     include ROXML
     # inner class for member: EmvDetails
@@ -1959,7 +1967,7 @@ end
     xml_accessor :transactionStatus
     xml_accessor :responseCode
     xml_accessor :responseReasonCode
-    xml_accessor :subscription
+    xml_accessor :subscription, :as => SubscriptionPaymentType
     xml_accessor :responseReasonDescription
     xml_accessor :authCode
     xml_accessor :aVSResponse
@@ -1990,8 +1998,9 @@ end
     xml_accessor :returnedItems, :as => ArrayOfReturnedItem
     xml_accessor :solution
     xml_accessor :emvDetails, :as => EmvDetails
+	xml_accessor :profile, :as => CustomerProfileIdType
   
-    def initialize(transId = nil, refTransId = nil, splitTenderId = nil, submitTimeUTC = nil, submitTimeLocal = nil, transactionType = nil, transactionStatus = nil, responseCode = nil, responseReasonCode = nil, subscription = nil, responseReasonDescription = nil, authCode = nil, aVSResponse = nil, cardCodeResponse = nil, cAVVResponse = nil, fDSFilterAction = nil, fDSFilters = nil, batch = nil, order = nil, requestedAmount = nil, authAmount = nil, settleAmount = nil, tax = nil, shipping = nil, duty = nil, lineItems = nil, prepaidBalanceRemaining = nil, taxExempt = nil, payment = nil, customer = nil, billTo = nil, shipTo = nil, recurringBilling = nil, customerIP = nil, product = nil, marketType = nil, mobileDeviceId = nil, returnedItems = nil, solution = nil, emvDetails = nil)
+    def initialize(transId = nil, refTransId = nil, splitTenderId = nil, submitTimeUTC = nil, submitTimeLocal = nil, transactionType = nil, transactionStatus = nil, responseCode = nil, responseReasonCode = nil, subscription = nil, responseReasonDescription = nil, authCode = nil, aVSResponse = nil, cardCodeResponse = nil, cAVVResponse = nil, fDSFilterAction = nil, fDSFilters = nil, batch = nil, order = nil, requestedAmount = nil, authAmount = nil, settleAmount = nil, tax = nil, shipping = nil, duty = nil, lineItems = nil, prepaidBalanceRemaining = nil, taxExempt = nil, payment = nil, customer = nil, billTo = nil, shipTo = nil, recurringBilling = nil, customerIP = nil, product = nil, marketType = nil, mobileDeviceId = nil, returnedItems = nil, solution = nil, emvDetails = nil, profile = nil)
       @transId = transId
       @refTransId = refTransId
       @splitTenderId = splitTenderId
@@ -2032,6 +2041,7 @@ end
       @returnedItems = returnedItems
       @solution = solution
       @emvDetails = emvDetails
+	  @profile = profile
     end
   end
   
@@ -2090,6 +2100,7 @@ end
   #   secureAcceptance - TransactionResponse::secureAcceptance
   #   emvResponse - TransactionResponse::emvResponse
   #   transHashSha2 - SOAP::SOAPString
+  #   profile - CustomerProfileIdType
   class TransactionResponse
     include ROXML
     # inner class for member: prePaidCard
@@ -2269,8 +2280,9 @@ end
     xml_accessor :secureAcceptance, :as => SecureAcceptance
     xml_accessor :emvResponse, :as => EmvResponse
     xml_accessor :transHashSha2
+	xml_accessor :profile, :as => CustomerProfileIdType
   
-    def initialize(responseCode = nil, rawResponseCode = nil, authCode = nil, avsResultCode = nil, cvvResultCode = nil, cavvResultCode = nil, transId = nil, refTransID = nil, transHash = nil, testRequest = nil, accountNumber = nil, accountType = nil, splitTenderId = nil, prePaidCard = nil, messages = nil, errors = nil, splitTenderPayments = nil, userFields = nil, shipTo = nil, secureAcceptance = nil, emvResponse = nil, transHashSha2 = nil)
+    def initialize(responseCode = nil, rawResponseCode = nil, authCode = nil, avsResultCode = nil, cvvResultCode = nil, cavvResultCode = nil, transId = nil, refTransID = nil, transHash = nil, testRequest = nil, accountNumber = nil, accountType = nil, splitTenderId = nil, prePaidCard = nil, messages = nil, errors = nil, splitTenderPayments = nil, userFields = nil, shipTo = nil, secureAcceptance = nil, emvResponse = nil, transHashSha2 = nil, profile = nil)
       @responseCode = responseCode
       @rawResponseCode = rawResponseCode
       @authCode = authCode
@@ -2293,6 +2305,7 @@ end
       @secureAcceptance = secureAcceptance
       @emvResponse = emvResponse
       @transHashSha2 = transHashSha2
+	  @profile = profile
     end
   end
   
@@ -2522,7 +2535,7 @@ end
   #   cardCode - (any)
   class PaymentProfile
     include ROXML
-    xml_accessor :paymentProfileId, :as => Fixnum
+    xml_accessor :paymentProfileId, :as => Integer
     xml_accessor :cardCode
   
     def initialize(paymentProfileId = nil, cardCode = nil)
@@ -2539,9 +2552,9 @@ end
   class CustomerProfilePaymentType
     include ROXML
     xml_accessor :createProfile
-    xml_accessor :customerProfileId, :as => Fixnum
+    xml_accessor :customerProfileId, :as => Integer
     xml_accessor :paymentProfile, :as => PaymentProfile
-    xml_accessor :shippingProfileId, :as => Fixnum
+    xml_accessor :shippingProfileId, :as => Integer
   
     def initialize(createProfile = nil, customerProfileId = nil, paymentProfile = nil, shippingProfileId = nil)
       @createProfile = createProfile
@@ -3141,18 +3154,24 @@ end
   #   merchantAuthentication - MerchantAuthenticationType
   #   refId - SOAP::SOAPString
   #   customerProfileId - (any)
+  #   merchantCustomerId - SOAP::SOAPString
+  #   email - SOAP::SOAPString
   #   unmaskExpirationDate - SOAP::SOAPBoolean
   class GetCustomerProfileRequest 
     include ROXML
     xml_accessor :merchantAuthentication
     xml_accessor :refId
     xml_accessor :customerProfileId
+	xml_accessor :merchantCustomerId
+	xml_accessor :email
     xml_accessor :unmaskExpirationDate
   
-    def initialize(merchantAuthentication = nil, refId = nil, customerProfileId = nil, unmaskExpirationDate = nil)
+    def initialize(merchantAuthentication = nil, refId = nil, customerProfileId = nil, unmaskExpirationDate = nil, merchantCustomerId = nil, email = nil)
       @merchantAuthentication = merchantAuthentication
       @refId = refId
       @customerProfileId = customerProfileId
+	  @merchantCustomerId = merchantCustomerId
+	  @email = email
       @unmaskExpirationDate = unmaskExpirationDate
     end
   end
@@ -3677,18 +3696,24 @@ end
   #   messages - MessagesType
   #   sessionToken - SOAP::SOAPString
   #   transaction - TransactionDetailsType
+  #   clientId - SOAP::SOAPString
+  #   transrefId - SOAP::SOAPString
   class GetTransactionDetailsResponse 
     include ROXML
     xml_accessor :refId
     xml_accessor :messages, :as => MessagesType
     xml_accessor :sessionToken
     xml_accessor :transaction, :as => TransactionDetailsType
+	xml_accessor :clientId
+	xml_accessor :transrefId
   
-    def initialize(refId = nil, messages = nil, sessionToken = nil, transaction = nil)
+    def initialize(refId = nil, messages = nil, sessionToken = nil, transaction = nil, clientId = nil, transrefId = nil)
       @refId = refId
       @messages = messages
       @sessionToken = sessionToken
       @transaction = transaction
+	  @clientId = clientId
+	  @transrefId = transrefId
     end
   end
   
@@ -3922,7 +3947,33 @@ end
       @totalNumInResultSet = totalNumInResultSet
     end
   end
+
+  # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}getTransactionListForCustomerRequest
+  #   merchantAuthentication - MerchantAuthenticationType
+  #   refId - SOAP::SOAPString
+  #   customerProfileId - SOAP::SOAPString
+  #   customerPaymentProfileId - SOAP::SOAPString
+  #	  sorting - TransactionListSorting
+  #   paging - Paging
+  class GetTransactionListForCustomerRequest 
+    include ROXML
+    xml_accessor :merchantAuthentication
+    xml_accessor :refId
+	xml_accessor :customerProfileId
+	xml_accessor :customerPaymentProfileId
+	xml_accessor :sorting, :as => TransactionListSorting
+    xml_accessor :paging, :as => Paging
   
+    def initialize(merchantAuthentication = nil, refId = nil, customerProfileId = nil, customerPaymentProfileId = nil, sorting = nil, paging = nil)
+      @merchantAuthentication = merchantAuthentication
+      @refId = refId
+	  @customerProfileId = customerProfileId
+	  @customerPaymentProfileId = customerPaymentProfileId
+	  @sorting = sorting
+	  @paging = paging
+    end
+  end
+
   # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}mobileDeviceRegistrationRequest
   #   merchantAuthentication - MerchantAuthenticationType
   #   refId - SOAP::SOAPString
@@ -4930,4 +4981,5 @@ end
       @token = token
     end
   end
+  
 end
