@@ -1,7 +1,8 @@
 module AuthorizeNet
   
   # The core, API agnostic transaction class. You shouldn't instantiate this one.
-  # Instead you should use AuthorizeNet::AIM::Transaction, AuthorizeNet::SIM::Transaction or AuthorizeNet::ARB::Transaction.
+  # Instead you should use AuthorizeNet::AIM::Transaction,
+  # AuthorizeNet::SIM::Transaction or AuthorizeNet::ARB::Transaction.
   class Transaction
     
     include AuthorizeNet::TypeConversions
@@ -12,16 +13,20 @@ module AuthorizeNet
     # Fields to convert to/from BigDecimal.
     @@decimal_fields = []
     
-    # DO NOT USE. Instantiate AuthorizeNet::AIM::Transaction, AuthorizeNet::SIM::Transaction or AuthorizeNet::ARB::Transaction instead.
+    # DO NOT USE. Instantiate AuthorizeNet::AIM::Transaction,
+    # AuthorizeNet::SIM::Transaction or AuthorizeNet::ARB::Transaction instead.
     def initialize()
       @fields ||= {}
     end
     
-    # Sets arbitrary API fields, overwriting existing values if they exist. Takes a hash of key/value pairs, 
-    # where the keys are the field names without the "x_" prefix. You can set a field to Nil to unset it. If
-    # the value is an array, each value in the array will be added. For example, set_fields({:line_item => 
-    # ["item1<|>golf balls<|><|>2<|>18.95<|>Y", "item2<|>golf bag<|>Wilson golf carry bag, red<|>1<|>39.99<|>"]})
-    # would generate two x_line_item fields in the transaction. One for each value in the array.
+    # Sets arbitrary API fields, overwriting existing values if they exist.
+    # Takes a hash of key/value pairs, where the keys are the field names
+    # without the "x_" prefix. You can set a field to Nil to unset it. If the
+    # value is an array, each value in the array will be added. For example,
+    # set_fields({:line_item => ["item1<|>golf balls<|><|>2<|>18.95<|>Y",
+    # "item2<|>golf bag<|>Wilson golf carry bag, red<|>1<|>39.99<|>"]})
+    # would generate two x_line_item fields in the transaction, one for
+    # each value in the array.
     def set_fields(fields = {})
       @fields.merge!(fields)
       @fields.reject! {|k, v| v.nil?}
@@ -38,7 +43,8 @@ module AuthorizeNet
       @fields.merge!(address.to_hash)
     end
     
-    # Takes an instance of AuthorizeNet::ShippingAddress and adds it to the transaction.
+    # Takes an instance of AuthorizeNet::ShippingAddress and adds it to the
+    # transaction.
     def set_shipping_address(address)
       @fields.merge!(address.to_hash)
     end
